requirements.txt
appdirs==1.4.0
Babel==2.3.4
Django==1.10
django-phonenumber-field==1.1.0
djangorestframework==3.5.3
packaging==16.8
phonenumberslite==8.2.0
psycopg2==2.6.2
pyparsing==2.1.10
pytz==2016.10
six==1.10.0
